# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'What to watch REST API'
  description: "Данный проект представляет собой REST API для управления базой данных фильмов,\n                        включая функции оставления комментариев, регистрации пользователей и управления\n                        списком избранных фильмов. Сервис интегрируется с КиноПоиском, что позволяет\n                        получать комментарии к фильмам и предоставляет пользователям возможность активно взаимодействовать с ними."
  version: 1.0.0
servers:
  -
    url: 'https://localhost'
paths:
...
  '/api/v1/genres/{genre_id}':
    put:
      summary: 'PATCH api/v1/genres/{genre_id}'
      operationId: pATCHApiv1genresgenreId
      description: 'Обновление жанра'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    name: Драма
                properties:
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: Драма
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
      tags:
        - Genre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Поле value не может превышать 32 символов.'
                  example: lvjzuiskwaradvwgzmvvsdja
              required:
                - name
    delete:
      summary: 'DELETE api/v1/genres/{genre_id}'
      operationId: dELETEApiv1genresgenreId
      description: 'Удаление жанра'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
      tags:
        - Genre
    parameters:
      -
        in: path
        name: genre_id
        description: 'Id жанра'
        example: 4
        required: true
        schema:
          type: integer
...
  '/api/v1/promo/{film_id}':
    post:
      summary: 'POST /api/v1/promo/{film_id}'
      operationId: pOSTapiv1promofilmId
      description: 'Установка фильма как промо'
      parameters: []
      responses:
        201:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Фильм с id 65 установлен как промо-фильм'
                properties:
                  message:
                    type: string
                    example: 'Фильм с id 65 установлен как промо-фильм'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
        409:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Фильм с id 65 уже установлен как промо-фильм'
                properties:
                  error:
                    type: string
                    example: 'Фильм с id 65 уже установлен как промо-фильм'
      tags:
        - Promo
    delete:
      summary: 'DELETE /api/v1/promo/{film_id}'
      operationId: dELETEapiv1promofilmId
      description: 'Удаление фильма из списка промо фильмов'
      parameters: []
      responses:
        204:
          description: ''
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет активной сессии'
                properties:
                  error:
                    type: string
                    example: 'Нет активной сессии'
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Нет доступа к дейсвтию!'
                properties:
                  error:
                    type: string
                    example: 'Нет доступа к дейсвтию!'
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  error: 'Запрашиваемая страница не существует'
                properties:
                  error:
                    type: string
                    example: 'Запрашиваемая страница не существует'
      tags:
        - Promo
    parameters:
      -
        in: path
        name: film_id
        description: 'The ID of the film.'
        example: 65
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: null
security:
  -
    default: []